fix(dbus): run hosted bus with role passwd jail
