+++ daemons.c
@@ -30,6 +30,7 @@
 
 #include "init.h"
 #include "daemons.h"
+#include "dbus_daemon.h"
 #include "tsh.h"
 #include "utils/fs.h"
 
@@ -41,20 +42,22 @@
 
 struct pv_init_daemon daemons[] = {
 	{ "hwrngd", 0, 1, "/usr/bin/rngd", "/usr/bin/rngd -f",
-	  DM_EMBEDDED | DM_STANDALONE, 0 },
+	  DM_EMBEDDED | DM_STANDALONE, NULL, 0 },
 #ifdef PANTAVISOR_XCONNECT
 	{ "pv-xconnect", 0, 1, "/usr/bin/pv-xconnect", "/usr/bin/pv-xconnect",
-	  DM_ALL, 0 },
+	  DM_ALL, NULL, 0 },
 #endif
 #ifdef PANTAVISOR_XCONNECT_DBUS_SYSTEMBUS
 	// No --system: that is shorthand for --config-file=.../system.conf and
 	// would conflict with our generated --config-file. Our config declares
-	// <type>system</type>, so this is still a system bus.
+	// <type>system</type>, so this is still a system bus. Spawned in a
+	// passwd jail so its policy can resolve the per-role masquerade uids
+	// without touching the rootfs /etc/passwd.
 	{ PV_DBUS_SYSTEMBUS_DAEMON, 0, 1, "/usr/bin/dbus-daemon",
 	  "/usr/bin/dbus-daemon --nofork --nopidfile --config-file=" PV_DBUS_SYSTEMBUS_CONF,
-	  DM_ALL, 0 },
+	  DM_ALL, PV_DBUS_SYSTEMBUS_PASSWD, 0 },
 #endif
-	{ 0, 0, 0, 0, 0, 0, 0 }
+	{ 0, 0, 0, 0, 0, 0, NULL, 0 }
 };
 
 static int daemon_spawn(struct pv_init_daemon *self)
@@ -70,8 +73,8 @@ static int daemon_spawn(struct pv_init_daemon *self)
 		char out_name[64], err_name[64];
 		snprintf(out_name, sizeof(out_name), "%s-out", self->name);
 		snprintf(err_name, sizeof(err_name), "%s-err", self->name);
-		self->pid =
-			tsh_run_daemon_logserver(self->cmd, out_name, err_name);
+		self->pid = tsh_run_daemon_logserver_jail(
+			self->cmd, out_name, err_name, self->jail_passwd);
 #else
 		self->pid = tsh_run(self->cmd, 0, 0);
 #endif
@@ -167,83 +170,12 @@ void pv_init_stop_daemons(void)
 	}
 }
 
-#ifdef PANTAVISOR_XCONNECT_DBUS_SYSTEMBUS
-// Default-deny base policy for the hosted system bus. Per-name allow rules are
-// generated by pantavisor into PV_DBUS_SYSTEMBUS_POLICYDIR from the owns/allow
-// declarations in container service manifests (see state.c). Rules are matched
-// in order, last match wins: method calls and name ownership are denied by
-// default and only re-granted by the generated per-uid fragments; calls to the
-// bus driver, replies and signals stay allowed so peers can connect and reply.
-static const char dbus_systembus_base_conf[] =
-	"<!DOCTYPE busconfig PUBLIC "
-	"\"-//freedesktop//DTD D-Bus Bus Configuration 1.0//EN\" "
-	"\"http://www.freedesktop.org/standards/dbus/1.0/busconfig.dtd\">\n"
-	"<busconfig>\n"
-	"  <type>system</type>\n"
-	"  <listen>unix:path=" PV_DBUS_SYSTEMBUS_SOCKET "</listen>\n"
-	"  <auth>EXTERNAL</auth>\n"
-	"\n"
-	"  <policy context=\"default\">\n"
-	"    <allow user=\"*\"/>\n"
-	"    <deny own=\"*\"/>\n"
-	"    <deny send_type=\"method_call\"/>\n"
-	"    <allow send_destination=\"org.freedesktop.DBus\"/>\n"
-	"    <allow receive_sender=\"org.freedesktop.DBus\"/>\n"
-	"    <allow send_requested_reply=\"true\" send_type=\"method_return\"/>\n"
-	"    <allow send_requested_reply=\"true\" send_type=\"error\"/>\n"
-	"    <allow receive_requested_reply=\"true\" receive_type=\"method_return\"/>\n"
-	"    <allow receive_requested_reply=\"true\" receive_type=\"error\"/>\n"
-	"    <allow send_type=\"signal\"/>\n"
-	"    <allow receive_type=\"signal\"/>\n"
-	"  </policy>\n"
-	"\n"
-	"  <includedir>" PV_DBUS_SYSTEMBUS_POLICYDIR "</includedir>\n"
-	"</busconfig>\n";
-
-// Gate the hosted dbus-daemon on the runtime config entry (defaults to true
-// when built in) and lay down its runtime dir + base config before spawn.
-static void dbus_systembus_prepare(void)
-{
-	if (!pv_config_get_bool(PV_XCONNECT_DBUS_SYSTEMBUS_ENABLED)) {
-		pv_log(INFO,
-		       "hosted dbus system bus disabled via config, not starting %s",
-		       PV_DBUS_SYSTEMBUS_DAEMON);
-		for (int i = 0; daemons[i].name; i++) {
-			if (!strcmp(daemons[i].name,
-				    PV_DBUS_SYSTEMBUS_DAEMON)) {
-				daemons[i].respawn = 0;
-				daemons[i].pid = -1;
-			}
-		}
-		return;
-	}
-
-	if (pv_fs_mkdir_p(PV_DBUS_SYSTEMBUS_POLICYDIR, 0755)) {
-		pv_log(ERROR, "could not create %s: %s",
-		       PV_DBUS_SYSTEMBUS_POLICYDIR, strerror(errno));
-		return;
-	}
-
-	FILE *f = fopen(PV_DBUS_SYSTEMBUS_CONF, "w");
-	if (!f) {
-		pv_log(ERROR, "could not write %s: %s", PV_DBUS_SYSTEMBUS_CONF,
-		       strerror(errno));
-		return;
-	}
-	fputs(dbus_systembus_base_conf, f);
-	fclose(f);
-
-	pv_log(INFO, "hosted dbus system bus enabled, base config at %s",
-	       PV_DBUS_SYSTEMBUS_CONF);
-}
-#endif
-
 static int pv_daemons_init(struct pv_init *this)
 {
 	init_mode_t mode = pv_config_get_system_init_mode();
 
 #ifdef PANTAVISOR_XCONNECT_DBUS_SYSTEMBUS
-	dbus_systembus_prepare();
+	pv_dbus_daemon_prepare();
 #endif
 
 	pv_init_spawn_daemons(mode);
